feat(autor): adiciona exception de delecao para autor

+++ src/main/java/com/example/LibraryAPI/services/AutorService.java
@@ -1,26 +1,26 @@
 package com.example.LibraryAPI.services;
 
+import com.example.LibraryAPI.exceptions.OperacaoNaoPermitida;
 import com.example.LibraryAPI.model.Autor;
 import com.example.LibraryAPI.repository.AutorRepository;
+import com.example.LibraryAPI.repository.LivroRepository;
 import com.example.LibraryAPI.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.lang.reflect.AnnotatedArrayType;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository autorRepository;
     private final AutorValidator validator;
-
-    @Autowired
-    public AutorService(AutorRepository autorRepository, AutorValidator validator) {
-        this.autorRepository = autorRepository;
-        this.validator = validator;
-    }
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor){
         validator.validar(autor);
@@ -40,6 +40,8 @@ public class AutorService {
     }
 
     public void deleteById(Autor autor){
+        if(possuiLivro(autor))
+            throw  new OperacaoNaoPermitida("Não eh permitido excluir autor possui livros cadastrados");
         autorRepository.delete(autor);
     }
 
@@ -57,5 +59,9 @@ public class AutorService {
         }
 
         return autorRepository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
